requirements.txt
pymongo==4.12.0
python-dotenv==1.0.0
psutil==5.9.8
docker==7.0.0
requests==2.31.0
pydantic==2.6.1
tqdm==4.66.1